refactor: folder structure
Placed all location based components in location folder and setup the conditional rendering logic for loading different questions (pages) on button click
