+++ app/components/postjob/location/JobLocation.jsx
@@ -1,10 +1,11 @@
 import React, { useState } from 'react'
-import { useData } from '../../post-job/PostJobContext';
+import { useData } from '../../../post-job/PostJobContext';
 import JobLocationDetails from "./JobLocationDetails";
 import PreferredCandidatesForm from "./PreferredCandidatesForm";
 import SelectJobLocation from "./SelectJobLocation";
 import SelectJobType from "./SelectJobType";
 import Input from './Input';
+import Button from "../Button";
 
 const LoadQuestion = ({jobType}) =>
 {
@@ -39,7 +40,7 @@ const LoadQuestion = ({jobType}) =>
 
 const JobLocation = ({setQuestionNumber}) => {
     const {state:{location } , dispatch} = useData();
-    const {jobTitle , jobType, showJobLocation, streetAddress, city, state, country, preferredLocations} = location;
+    const {jobTitle , jobType, showJobLocation, streetAddress, city, state, country} = location;
 
     const isDisabled = () =>
     {
@@ -57,7 +58,7 @@ const JobLocation = ({setQuestionNumber}) => {
         <Input id="jobTitle" value={jobTitle} changeHandler={(inputField,userInput)=>dispatch({type:"LOCATION_FIELDS" , payload:{inputField, userInput} })} type={"text"} placeholder={"Job title (maximum 160 Characters)"} label="Job Title" required={true}/>
         <SelectJobType />
         <LoadQuestion jobType={jobType}/>
-        <button className="px-4 py-3 bg-white-light rounded-xl text-sm font-bold w-full text-primary disabled:brightness-50" disabled={isDisabled()} onClick={()=>setQuestionNumber(2)}>Continue</button>
+        <Button isDisabled={isDisabled} clickHandler={()=>setQuestionNumber(2)} />
     </>
   )
 }
