+++ app/post-job/page.jsx
@@ -1,14 +1,16 @@
 'use client';
-import React, { useState } from 'react'
+import React, { useState , useEffect } from 'react'
 import { PostJobProvider } from "./PostJobContext";
 
 import ProgressBar from "../components/ProgressBar";
-import JobLocation from "../components/postjob/JobLocation";
+import JobLocation from "../components/postjob/location/JobLocation";
 import Salary from "../components/postjob/salary/Salary";
 
 const page = () => {
-    const [step, setStep] = useState(1);
-    const [questionNumber, setQuestionNumber] = useState(1)
+  const [step, setStep] = useState(1);
+  const [questionNumber, setQuestionNumber] = useState(1)
+
+  useEffect(()=>{window.scrollTo(0, 0)},[questionNumber])
 
   return (
     <div className=" my-4 mx-6 space-y-4">
@@ -24,8 +26,8 @@ const page = () => {
             <span className="text-sm text-required">*Marked fields are mandatory</span>
         </header>
         <PostJobProvider>
-          {/* <JobLocation setQuestionNumber={setQuestionNumber}/> */}
-          <Salary />
+          {questionNumber==1 && <JobLocation setQuestionNumber={setQuestionNumber}/>}
+          {questionNumber==2 && <Salary setQuestionNumber={setQuestionNumber}/>}
         </PostJobProvider>
       </main>
     </div>
